% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pred.R
\name{.blp_pred_single}
\alias{.blp_pred_single}
\title{Internal function to compute BLP for prediction (single repetition)}
\usage{
.blp_pred_single(Y, predicted_y, cluster_id = NULL)
}
\arguments{
\item{Y}{Outcome vector}

\item{predicted_y}{Predicted Y vector for this repetition}

\item{cluster_id}{Optional vector of cluster identifiers for cluster-robust SEs.}
}
\value{
data.table with BLP estimates
}
\description{
Internal function to compute BLP for prediction (single repetition)
}
\keyword{internal}
