% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_restricted.R
\name{.calc_se_diff_regs}
\alias{.calc_se_diff_regs}
\title{Compute SE of difference between two regression coefficients}
\usage{
.calc_se_diff_regs(reg1, reg2, coef_name)
}
\arguments{
\item{reg1}{Result from reg_from_formula (list with coef, model_matrix, residuals)}

\item{reg2}{Result from reg_from_formula (list with coef, model_matrix, residuals)}

\item{coef_name}{Name of the coefficient to compare}
}
\value{
Named vector with 'estimate' and 'se'
}
\description{
Computes the estimate and standard error of the difference between
corresponding coefficients from two regressions (reg1 - reg2).
}
\keyword{internal}
