% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hte.R
\name{.compute_coef_diff}
\alias{.compute_coef_diff}
\title{Compute difference between two coefficients with proper SE}
\usage{
.compute_coef_diff(coef, vcov, coef1, coef2)
}
\arguments{
\item{coef}{Coefficient matrix from regression}

\item{vcov}{Variance-covariance matrix}

\item{coef1}{Name of first coefficient}

\item{coef2}{Name of second coefficient}
}
\value{
data.table with estimate and se
}
\description{
Compute difference between two coefficients with proper SE
}
\keyword{internal}
