% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_training.R
\name{.encode_factors_if_needed}
\alias{.encode_factors_if_needed}
\title{Encode factor/character columns as dummy variables}
\usage{
.encode_factors_if_needed(.X)
}
\arguments{
\item{.X}{data.frame of covariates (may contain factor/character columns)}
}
\value{
data.frame with factor/character columns replaced by dummy variables
}
\description{
Used internally for algorithms that don't support factor features natively.
}
\keyword{internal}
