% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_training.R
\name{.fit_model_grf}
\alias{.fit_model_grf}
\title{Fit a model using grf (generalized random forest)}
\usage{
.fit_model_grf(.Y, .X, task_type = c("regr", "classif"), weights = NULL)
}
\description{
Fit a model using grf (generalized random forest)
}
\keyword{internal}
