% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_training.R
\name{.predict_model}
\alias{.predict_model}
\title{Generic prediction helper for both mlr3 and grf models}
\usage{
.predict_model(model, newdata)
}
\description{
Generic prediction helper for both mlr3 and grf models
}
\keyword{internal}
