% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleHTE-package.R
\docType{package}
\name{ensembleHTE-package}
\alias{ensembleHTE}
\alias{ensembleHTE-package}
\title{ensembleHTE: Ensemble Methods for Heterogeneous Treatment Effect Estimation}
\description{
The ensembleHTE package provides tools for estimating heterogeneous treatment
effects (HTE) and general predictions using ensemble learning methods. It combines
multiple machine learning algorithms with repeated cross-fitting to improve
statistical power.
}
\details{
The package provides two main estimation functions:
\itemize{
  \item \code{\link{ensemble_hte}}: Estimates individual treatment effects (ITE)
    using metalearner strategies (R-learner, T-learner, S-learner, X-learner)
  \item \code{\link{ensemble_pred}}: Estimates outcome predictions using an
    ensemble of ML algorithms
}

Key features:
\itemize{
  \item Ensemble learning combining multiple ML algorithms (grf, lm, ranger, etc.)
  \item Repeated K-fold cross-fitting (M repetitions) for improved power
  \item Cross-validated Best Linear Predictor (BLP) weights for ensemble combination
  \item Analysis tools: GATES, BLP, CLAN, GAVS for characterizing treatment effects
}
}
\section{Main functions}{

\itemize{
  \item \code{\link{ensemble_hte}}: Fit ensemble HTE model with metalearners
  \item \code{\link{ensemble_pred}}: Fit ensemble prediction model
  \item \code{\link{gates}}: Group Average Treatment Effects
  \item \code{\link{blp}}: Best Linear Predictor of CATE
  \item \code{\link{clan}}: Classification Analysis by covariate
  \item \code{\link{gavs}}: Group Averages
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bfava/ensembleHTE-dev}
  \item Report bugs at \url{https://github.com/bfava/ensembleHTE-dev/issues}
}

}
\author{
\strong{Maintainer}: Bruno Fava \email{brunovnfava@gmail.com}

}
\keyword{internal}
