% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stars}
\alias{get_stars}
\title{Get Significance Stars}
\usage{
get_stars(p)
}
\arguments{
\item{p}{Numeric vector of p-values}
}
\value{
Character vector of significance stars
}
\description{
Converts p-values to significance stars following standard convention:
\itemize{
  \item \code{***}: p < 0.001
  \item \code{**}: p < 0.01
  \item \code{*}: p < 0.05
  \item \code{.}: p < 0.1
  \item (empty): p >= 0.1
}
}
\examples{
\dontrun{
get_stars(c(0.0001, 0.005, 0.03, 0.08, 0.5))
# Returns: "***" "**" "*" "." ""
}

}
\keyword{internal}
