% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microcredit_covariates}
\alias{microcredit_covariates}
\title{Control Covariates for the Microcredit Dataset}
\format{
A character vector of length 24 containing column names from the
  \code{\link{microcredit}} dataset:
  \code{"css_creditscorefinal"}, \code{"css_yearsataddress"},
  \code{"own_house"}, \code{"sari"}, \code{"own_anybus"},
  \code{"max_yearsinbusiness"}, \code{"css_regularworkers"},
  \code{"css_traveltime"}, \code{"css_travelcost"},
  \code{"css_stockvalue"}, \code{"css_assetvalue"},
  \code{"css_noofbusiness"}, \code{"hhsize"}, \code{"age"},
  \code{"gender"}, \code{"education"}, \code{"married"},
  \code{"savingsamt"}, \code{"hhinc_yrly_base"},
  \code{"profits_yrly_base"}, \code{"rev_yrly_base"},
  \code{"exp_yrly_base"}, \code{"hhexp_yrly_base"},
  \code{"hhasset_yrly_base"}.
}
\source{
Karlan, D. and Zinman, J. (2011). Microcredit in Theory and Practice: Using
Randomized Credit Scoring for Impact Evaluation. \emph{Science},
\strong{332}(6035), 1278--1284.
\doi{10.1126/science.1200138}

Athey, S., Fava, B., Karlan, D., Osman, A. and Zinman, J. (2025). Profits
and Social Impacts: Complements vs. Tradeoffs for Lenders in Three
Countries. Working paper.
}
\usage{
microcredit_covariates
}
\description{
A character vector of 24 covariate names used as controls in the analysis
of the Philippine microcredit experiment by Athey, Fava, Karlan, Osman &
Zinman (2025). Provided as a convenience for use in formulas and subsetting.
}
\examples{
data(microcredit_covariates)
microcredit_covariates

# Use to subset the microcredit data
data(microcredit)
dat <- microcredit[, c("hhinc_yrly_end", "treat", microcredit_covariates)]
}
\seealso{
\code{\link{microcredit}}
}
\keyword{datasets}
