% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_training.R
\name{predict_newdata.grf_wrapper}
\alias{predict_newdata.grf_wrapper}
\title{Predict method for grf_wrapper}
\usage{
predict_newdata.grf_wrapper(object, newdata)
}
\description{
Predict method for grf_wrapper
}
\keyword{internal}
