% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pred.R
\name{print.blp_pred_results}
\alias{print.blp_pred_results}
\title{Print method for blp_pred_results objects}
\usage{
\method{print}{blp_pred_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{blp_pred_results} from \code{blp_pred()}}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for blp_pred_results objects
}
