% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hte.R
\name{print.blp_results}
\alias{print.blp_results}
\title{Print method for blp_results objects}
\usage{
\method{print}{blp_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{blp_results} from \code{blp()}}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for blp_results objects
}
