% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_restricted.R
\name{print.gates_compare_results}
\alias{print.gates_compare_results}
\title{Print method for gates_compare_results objects}
\usage{
\method{print}{gates_compare_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gates_compare_results} from \code{gates_compare()}}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for gates_compare_results objects
}
