% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_restricted.R
\name{print.gates_restricted_results}
\alias{print.gates_restricted_results}
\title{Print method for gates_restricted_results objects}
\usage{
\method{print}{gates_restricted_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gates_restricted_results} from \code{gates_restricted()}}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for gates_restricted_results objects
}
