% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hte.R
\name{print.gates_results}
\alias{print.gates_results}
\title{Print method for gates_results objects}
\usage{
\method{print}{gates_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gates_results} from \code{gates()}}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for gates_results objects
}
