% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_restricted.R
\name{print.gavs_compare_results}
\alias{print.gavs_compare_results}
\title{Print method for gavs_compare_results objects}
\usage{
\method{print}{gavs_compare_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gavs_compare_results} from \code{gavs_compare()}}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for gavs_compare_results objects
}
