% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_restricted.R
\name{print.gavs_restricted_results}
\alias{print.gavs_restricted_results}
\title{Print method for gavs_restricted_results objects}
\usage{
\method{print}{gavs_restricted_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gavs_restricted_results} from \code{gavs_restricted()}}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for gavs_restricted_results objects
}
