% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hte.R
\name{print.gavs_results}
\alias{print.gavs_results}
\title{Print method for gavs_results objects}
\usage{
\method{print}{gavs_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gavs_results} from \code{gavs()}}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for gavs_results objects
}
