% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_training.R
\name{.fit_model_mlr3}
\alias{.fit_model_mlr3}
\title{Fit a model using mlr3}
\usage{
.fit_model_mlr3(
  .Y,
  .X,
  algorithm,
  task_type = c("regr", "classif"),
  tune = FALSE,
  tune_params = list(time = 30, cv_folds = 3, stagnation_iters = 250,
    stagnation_threshold = 0.01, measure = NULL),
  weights = NULL
)
}
\description{
Fit a model using mlr3
}
\keyword{internal}
