% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_training.R
\name{fit_model}
\alias{fit_model}
\title{Fit a model using the appropriate framework}
\usage{
fit_model(
  .Y,
  .X,
  algorithm,
  task_type = c("regr", "classif"),
  tune = FALSE,
  tune_params = list(time = 30, cv_folds = 3, stagnation_iters = 250,
    stagnation_threshold = 0.01, measure = NULL),
  weights = NULL
)
}
\description{
Dispatcher function that routes to mlr3 or grf based on algorithm
}
\keyword{internal}
