% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_hte.R
\name{print.clan_results}
\alias{print.clan_results}
\title{Print method for clan_results objects}
\usage{
\method{print}{clan_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{clan_results} from \code{clan()}}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for clan_results objects
}
