% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_common_inputs}
\alias{validate_common_inputs}
\title{Validate Common Inputs}
\usage{
validate_common_inputs(M, K, algorithms, ensemble_folds, n_cores)
}
\arguments{
\item{M}{Number of repetitions}

\item{K}{Number of folds}

\item{algorithms}{Vector of algorithm names}

\item{ensemble_folds}{Number of ensemble folds}

\item{n_cores}{Number of cores}
}
\value{
NULL if valid, raises error otherwise
}
\description{
Validates common input parameters for ensemble functions.
Internal function.
}
\keyword{internal}
